[REFACTORING] Split IMAPServerIDCommandTest
