+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerIDCommandTest.java
@@ -1,0 +1,78 @@
+/****************************************************************
+ * Licensed to the Apache Software Foundation (ASF) under one   *
+ * or more contributor license agreements.  See the NOTICE file *
+ * distributed with this work for additional information        *
+ * regarding copyright ownership.  The ASF licenses this file   *
+ * to you under the Apache License, Version 2.0 (the            *
+ * "License"); you may not use this file except in compliance   *
+ * with the License.  You may obtain a copy of the License at   *
+ *                                                              *
+ *   http://www.apache.org/licenses/LICENSE-2.0                 *
+ *                                                              *
+ * Unless required by applicable law or agreed to in writing,   *
+ * software distributed under the License is distributed on an  *
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY       *
+ * KIND, either express or implied.  See the License for the    *
+ * specific language governing permissions and limitations      *
+ * under the License.                                           *
+ ****************************************************************/
+
+package org.apache.james.imapserver.netty;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.Duration;
+
+import org.apache.james.util.concurrency.ConcurrentTestRunner;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.Test;
+
+
+@SuppressWarnings("checkstyle:membername")
+class IMAPServerIDCommandTest extends AbstractIMAPServerTest {
+    IMAPServer imapServer;
+
+    @AfterEach
+    void tearDown() {
+        if (imapServer != null) {
+            imapServer.destroy();
+        }
+    }
+
+    @Test
+    void idCommandShouldReturnNILWhenNoConfigured() throws Exception {
+        imapServer = createImapServer("imapServer.xml");
+
+        assertThat(
+            testIMAPClient.connect("127.0.0.1", imapServer.getListenAddresses().getFirst().getPort())
+                .sendCommand("ID (\"name\" \"Apache James\")"))
+            .contains("* ID NIL")
+            .contains("OK ID completed.");
+    }
+
+    @Test
+    void idCommandShouldReturnConfiguredResponse() throws Exception {
+        imapServer = createImapServer("imapServerIdCommandResponseFields.xml");
+        assertThat(
+            testIMAPClient.connect("127.0.0.1", imapServer.getListenAddresses().getFirst().getPort())
+                .sendCommand("ID (\"name\" \"Apache James\")"))
+            .contains("* ID (\"name\" \"Apache James\" \"version\" \"3.9.0\")")
+            .contains("OK ID completed.");
+    }
+
+    @Test
+    void concurrentIdCommandsInTheSameSessionShouldSucceed() throws Exception {
+        imapServer = createImapServer("imapServer.xml");
+
+        testIMAPClient.connect("127.0.0.1", imapServer.getListenAddresses().getFirst().getPort());
+        ConcurrentTestRunner.builder()
+            .operation((threadNumber, step) -> {
+                assertThat(testIMAPClient.sendCommand("ID (\"name\" \"Apache James\")"))
+                    .contains("* ID NIL")
+                    .contains("OK ID completed.");
+            })
+            .threadCount(20)
+            .operationCount(1)
+            .runSuccessfullyWithin(Duration.ofMinutes(5));
+    }
+}
